+++ components/src/main/java/uk/gov/hmrc/components/molecule/donut/DonutChartView.kt
@@ -65,6 +65,7 @@ class DonutChartView @JvmOverloads constructor(
     var value3Percent: Float = 0.0f
 
     init {
+        clearValues()
         context.theme.obtainStyledAttributes(attrs, R.styleable.DonutChartView, 0, 0).apply {
             try {
                 color1 = getColor(
@@ -95,6 +96,20 @@ class DonutChartView @JvmOverloads constructor(
                 recycle()
             }
         }
+    }
+
+    private fun clearValues() {
+        value1 = 0.0f
+        value2 = 0.0f
+        value3 = 0.0f
+
+        value1Percent = 0.0f
+        value2Percent = 0.0f
+        value3Percent = 0.0f
+
+        value1SweepAngle = 0f
+        value2SweepAngle = 0f
+        value3SweepAngle = 0f
     }
 
     private fun stripesOrPlain(stripes: Boolean, color: Int): Paint {
@@ -186,6 +201,7 @@ class DonutChartView @JvmOverloads constructor(
     }
 
     fun setValues(value1: Float, value2: Float, value3: Float = 0f, shouldAnimate: Boolean) {
+        clearValues()
         this@DonutChartView.value1 = value1
         this@DonutChartView.value2 = value2
         this@DonutChartView.value3 = value3
@@ -269,9 +285,17 @@ class DonutChartView @JvmOverloads constructor(
 
     private fun Animation.setOnAnimationEnd(onEnd: () -> Unit) {
         this.setAnimationListener(object : Animation.AnimationListener {
-            @Suppress("EmptyFunctionBlock") override fun onAnimationRepeat(animation: Animation?) {}
-            @Suppress("EmptyFunctionBlock") override fun onAnimationStart(animation: Animation?) {}
-            override fun onAnimationEnd(animation: Animation?) { onEnd() }
+            @Suppress("EmptyFunctionBlock")
+            override fun onAnimationRepeat(animation: Animation?) {
+            }
+
+            @Suppress("EmptyFunctionBlock")
+            override fun onAnimationStart(animation: Animation?) {
+            }
+
+            override fun onAnimationEnd(animation: Animation?) {
+                onEnd()
+            }
         })
     }